Development/Ecom-Microservice/auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
python-multipart
bcrypt
email-validator
psycopg2-binary
jinja2
authlib
python-dotenv
httpx
google-auth
google-auth-oauthlib
requests-oauthlib
requests
alembic
jwt
pydantic[email]
email-validator